Finding networks now works. Supremely ghetto solution to actually storing them, which we may need to change.
When findNetwork completes a network and returns it, it’s received as
null in findAllNetworks regardless of what is returned in findNetwork.
As a duct tape solution, I just stored all the completed networks in an
instance variable for now.

Left behind all the prints for future debugging. We can remove them all
when ready to submit.